Move metrics and shared analysis tools into snorkel/analysis
