+++ snorkel/analysis/metrics.py
@@ -1,8 +1,6 @@
 import numpy as np
-import sklearn.metrics as skm
-import torch
 
-from .utils import arraylike_to_numpy, pred_to_prob
+from .utils import arraylike_to_numpy
 
 
 def accuracy_score(gold, pred, prob=None, ignore_in_gold=[], ignore_in_pred=[]):
@@ -151,34 +149,6 @@ def f1_score(gold, pred, prob=None, **kwargs):
     return fbeta_score(gold, pred, prob, beta=1.0, **kwargs)
 
 
-def roc_auc_score(gold, pred, prob, ignore_in_gold=[], ignore_in_pred=[]):
-    """Compute the ROC AUC score, given the gold labels and predicted probs.
-
-    Args:
-        gold: A 1d array-like of gold labels
-        pred: Unused (kept for compatibility with expected metric func signature)
-        prob: A 2d array-like of predicted probabilities
-        ignore_in_gold: A list of labels for which elements having that gold
-            label will be ignored.
-
-    Returns:
-        roc_auc_score: The (float) roc_auc score
-    """
-    gold = arraylike_to_numpy(gold)
-
-    # Filter out the ignore_in_gold (but not ignore_in_pred)
-    # Note the current sub-functions (below) do not handle this...
-    if len(ignore_in_pred) > 0:
-        raise ValueError("ignore_in_pred not defined for ROC-AUC score.")
-    keep = [x not in ignore_in_gold for x in gold]
-    gold = gold[keep]
-    prob = prob[keep, :]
-
-    # Convert gold to one-hot indicator format, using the k inferred from probs
-    gold_s = pred_to_prob(torch.from_numpy(gold), k=prob.shape[1]).numpy()
-    return skm.roc_auc_score(gold_s, prob)
-
-
 def _get_mask(gold, pred, ignore_in_gold, ignore_in_pred):
     """Remove from gold, pred all items with labels designated to ignore."""
     mask = np.ones_like(gold).astype(bool)
@@ -208,7 +178,6 @@ METRICS = {
     "recall": recall_score,
     "f1": f1_score,
     "fbeta": fbeta_score,
-    "roc-auc": roc_auc_score,
 }
