+++ test/labeling/test_analysis.py
@@ -5,8 +5,6 @@ import pandas as pd
 import scipy.sparse as sparse
 
 from snorkel.labeling.analysis import (
-    confusion_matrix,
-    error_buckets,
     label_conflict,
     label_coverage,
     label_overlap,
@@ -133,40 +131,3 @@ class TestAnalysis(unittest.TestCase):
             }
         )
         pd.testing.assert_frame_equal(df.round(6), df_expected.round(6))
-
-    def test_error_buckets(self) -> None:
-        pred = [2, 1, 3, 1, 1, 3]
-        buckets = error_buckets(self.Y, pred, X=None)
-        self.assertEqual(
-            buckets, {(2, 1): [0], (1, 2): [1, 4], (3, 3): [2, 5], (1, 1): [3]}
-        )
-
-    def test_confusion_matrix(self) -> None:
-        pred = [0, 2, 2, 3, 1, 0, 1, 3]
-        gold = [1, 2, 2, 0, 3, 0, 2, 3]
-
-        mat = confusion_matrix(
-            gold, pred, null_pred=False, null_gold=False, normalize=False
-        )
-        mat_expected = np.array([[0, 1, 1], [0, 2, 0], [0, 0, 1]])
-        np.testing.assert_array_equal(mat, mat_expected)
-
-        mat = confusion_matrix(
-            gold, pred, null_pred=False, null_gold=False, normalize=True
-        )
-        mat_expected = np.array([[0, 1 / 8, 1 / 8], [0, 2 / 8, 0], [0, 0, 1 / 8]])
-        np.testing.assert_array_equal(mat, mat_expected)
-
-        mat = confusion_matrix(
-            gold, pred, null_pred=True, null_gold=False, normalize=False
-        )
-        mat_expected = np.array([[1, 0, 0], [0, 1, 1], [0, 2, 0], [0, 0, 1]])
-        np.testing.assert_array_equal(mat, mat_expected)
-
-        mat = confusion_matrix(
-            gold, pred, null_pred=True, null_gold=True, normalize=False
-        )
-        mat_expected = np.array(
-            [[1, 1, 0, 0], [0, 0, 1, 1], [0, 0, 2, 0], [1, 0, 0, 1]]
-        )
-        np.testing.assert_array_equal(mat, mat_expected)
